+++ server/src-lib/Hasura/GraphQL/Schema/Remote.hs
@@ -216,10 +216,12 @@ exact same value as the input (if perhaps represented differently); by discardin
 just forwarding the input, we avoid expanding variables if no preset needs be inserted.
 -}
 
--- | Helper, used to track whether an input value was altered during its parsing. At time of
--- writing, the only possible source of alteration is preset values. They might force evaluation of
--- variables, and encapsulation of sub-JSON expressions as new variables. Each parser indicates
--- whether such alteration took place within its part of the tree.
+-- | Helper, used to track whether an input value was altered during its parsing.
+-- There are two possible sources of alteration:
+--   - preset values, and
+--   - type name customizations.
+-- They might force evaluation of variables, and encapsulation of sub-JSON expressions as new variables.
+-- Each parser indicates whether such alteration took place within its part of the tree.
 -- See Note [Variable expansion in remote schema input parsers] for more information.
 newtype Altered = Altered {getAltered :: Bool}
   deriving (Show)
@@ -340,10 +342,12 @@ remoteFieldScalarParser customizeTypename (G.ScalarTypeDefinition description na
     { pType = schemaType,
       pParser = \case
         JSONValue v ->
+          -- Disallow short-circuit optimisation if the type name has been changed by remote schema customization
           pure $ (Altered $ G.getBaseType gType /= name, G.VVariable $ RemoteJSONValue (mkRemoteGType gType) v)
         GraphQLValue v -> case v of
           G.VVariable var -> do
             P.typeCheck False gType var
+            -- Disallow short-circuit optimisation if the type name has been changed by remote schema customization
             pure $ (Altered $ G.getBaseType (vType var) /= name, G.VVariable $ QueryVariable var {vType = mkRemoteGType (vType var)})
           _ -> pure (Altered False, QueryVariable <$> v)
     }
@@ -414,7 +418,10 @@ remoteInputObjectParser schemaDoc defn@(G.InputObjectTypeDefinition desc name _ 
 
       Right <$> P.memoizeOn 'remoteInputObjectParser defn do
         typename <- mkTypename name
-        argsParser <- argumentsParser valueDefns schemaDoc
+
+        -- Disallow short-circuit optimisation if the type name has been changed by remote schema customization
+        let altered = Altered $ typename /= name
+        argsParser <- fmap (first (<> altered)) <$> argumentsParser valueDefns schemaDoc
         pure $ fmap G.VObject <$> P.object typename desc argsParser
 
 -- | Variable expansion optimization.
@@ -433,7 +440,7 @@ shortCircuitIfUnaltered parser =
         result <- P.pParser parser value
         pure $ case result of
           -- The parser did yield a value, and it was unmodified by presets
-          -- we can short-citcuit by transforming the input value, therefore
+          -- we can short-circuit by transforming the input value, therefore
           -- "unpeeling" variables and avoiding extraneous JSON variables.
           Just (Altered False, _) -> Just $
             (Altered False,) $ case castWith (P.inputParserInput @k) value of
@@ -444,7 +451,7 @@ shortCircuitIfUnaltered parser =
               -- all the leaves of said value each be their own distinct value.
               JSONValue v -> G.VVariable $ RemoteJSONValue (toGraphQLType $ P.pType parser) v
           -- Otherwise either the parser did not yield any value, or a value
-          -- that has been altered by presets and permissions; we forward it
+          -- that has been altered by presets, permissions or type name customization; we forward it
           -- unoptimized.
           _ -> result
     }